JWT validation close #4
