+++ pkg/middleware/fiber_middleware.go
@@ -1,9 +1,12 @@
 package middleware
 
 import (
+	"os"
+
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/fiber/v2/middleware/cors"
 	"github.com/gofiber/fiber/v2/middleware/logger"
+	jwtware "github.com/gofiber/jwt/v3"
 )
 
 func FiberMiddleware(a *fiber.App) {
@@ -11,4 +14,10 @@ func FiberMiddleware(a *fiber.App) {
 		cors.New(),
 		logger.New(),
 	)
+
+	secret := os.Getenv("JWT_SECRET_KEY") // utils.GoDotEnvVariable("JWT_SECRET_KEY") //
+	a.Use("/api/v1/contacts", jwtware.New(jwtware.Config{
+		SigningKey: []byte(secret),
+		ContextKey: "auth",
+	}))
 }
