+++ pkg/utils/jwt.go
@@ -1,0 +1,42 @@
+package utils
+
+import (
+	"github.com/gofiber/fiber/v2"
+	"github.com/golang-jwt/jwt/v4"
+)
+
+func GetClaim(c *fiber.Ctx, key string) string {
+	auth := c.Locals("auth").(*jwt.Token)
+	claims := auth.Claims.(jwt.MapClaims)
+	value := claims[key].(string)
+	return value
+}
+
+/*
+func extractToken(c *fiber.Ctx) string {
+	bearToken := c.Get("Authorization")
+
+	// Normally Authorization HTTP header.
+	onlyToken := strings.Split(bearToken, " ")
+	if len(onlyToken) == 2 {
+		return onlyToken[1]
+	}
+
+	return ""
+}
+
+func verifyToken(c *fiber.Ctx) (*jwt.Token, error) {
+	tokenString := extractToken(c)
+
+	token, err := jwt.Parse(tokenString, jwtKeyFunc)
+	if err != nil {
+		return nil, err
+	}
+
+	return token, nil
+}
+
+func jwtKeyFunc(token *jwt.Token) (interface{}, error) {
+	return []byte(os.Getenv("JWT_SECRET_KEY")), nil
+}
+*/
