--- pkg/utils/jwt_decoder.go
@@ -1,67 +1,0 @@
-package utils
-
-import (
-	"os"
-	"strings"
-
-	"github.com/gofiber/fiber/v2"
-	"github.com/golang-jwt/jwt/v4"
-)
-
-// TokenMetadata struct to describe metadata in JWT.
-type TokenMetadata struct {
-	UserID  string
-	Expires int64
-}
-
-// ExtractTokenMetadata func to extract metadata from JWT.
-func ExtractTokenMetadata(c *fiber.Ctx) (*TokenMetadata, error) {
-	token, err := verifyToken(c)
-	if err != nil {
-		return nil, err
-	}
-
-	// Setting and checking token and credentials.
-	claims, ok := token.Claims.(jwt.MapClaims)
-	if ok && token.Valid {
-		// User ID.
-		userID := claims["id"].(string)
-
-		// Expires time.
-		expires := int64(claims["expires"].(float64))
-
-		return &TokenMetadata{
-			UserID:  userID,
-			Expires: expires,
-		}, nil
-	}
-
-	return nil, err
-}
-
-func extractToken(c *fiber.Ctx) string {
-	bearToken := c.Get("Authorization")
-
-	// Normally Authorization HTTP header.
-	onlyToken := strings.Split(bearToken, " ")
-	if len(onlyToken) == 2 {
-		return onlyToken[1]
-	}
-
-	return ""
-}
-
-func verifyToken(c *fiber.Ctx) (*jwt.Token, error) {
-	tokenString := extractToken(c)
-
-	token, err := jwt.Parse(tokenString, jwtKeyFunc)
-	if err != nil {
-		return nil, err
-	}
-
-	return token, nil
-}
-
-func jwtKeyFunc(token *jwt.Token) (interface{}, error) {
-	return []byte(os.Getenv("JWT_SECRET_KEY")), nil
-}
